Struct gmail1::UserMethodsBuilder

pub struct UserMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on user resources. It is not used directly, but through the Gmail hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate gmail1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gmail1::Gmail;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Gmail::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `drafts_create(...)`, `drafts_delete(...)`, `drafts_get(...)`, `drafts_list(...)`, `drafts_send(...)`, `drafts_update(...)`, `get_profile(...)`, `history_list(...)`, `labels_create(...)`, `labels_delete(...)`, `labels_get(...)`, `labels_list(...)`, `labels_patch(...)`, `labels_update(...)`, `messages_attachments_get(...)`, `messages_delete(...)`, `messages_get(...)`, `messages_import(...)`, `messages_insert(...)`, `messages_list(...)`, `messages_modify(...)`, `messages_send(...)`, `messages_trash(...)`, `messages_untrash(...)`, `threads_delete(...)`, `threads_get(...)`, `threads_list(...)`, `threads_modify(...)`, `threads_trash(...)` and `threads_untrash(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, NC, A> UserMethodsBuilder<'a, C, NC, A>

fn messages_import(&self, request: &Message, user_id: &str) -> UserMessageImportMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Imports a message into only this user's mailbox, with standard email delivery scanning and classification similar to receiving via SMTP. Does not send a message.

fn history_list(&self, user_id: &str) -> UserHistoryListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists the history of all changes to the given mailbox. History results are returned in chronological order (increasing historyId).

fn drafts_create(&self, request: &Draft, user_id: &str) -> UserDraftCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a new draft with the DRAFT label.

fn labels_create(&self, request: &Label, user_id: &str) -> UserLabelCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a new label.

fn labels_delete(&self, user_id: &str, id: &str) -> UserLabelDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified label and removes it from any messages and threads that it is applied to.

fn labels_get(&self, user_id: &str, id: &str) -> UserLabelGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the specified label.

fn messages_trash(&self, user_id: &str, id: &str) -> UserMessageTrashMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Moves the specified message to the trash.

fn drafts_send(&self, request: &Draft, user_id: &str) -> UserDraftSendMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sends the specified, existing draft to the recipients in the To, Cc, and Bcc headers.

fn messages_untrash(&self, user_id: &str, id: &str) -> UserMessageUntrashMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes the specified message from the trash.

fn labels_list(&self, user_id: &str) -> UserLabelListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all labels in the user's mailbox.

fn messages_delete(&self, user_id: &str, id: &str) -> UserMessageDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified message. This operation cannot be undone. Prefer messages.trash instead.

fn drafts_update(&self, request: &Draft, user_id: &str, id: &str) -> UserDraftUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Replaces a draft's content.

fn drafts_get(&self, user_id: &str, id: &str) -> UserDraftGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the specified draft.

fn labels_update(&self, request: &Label, user_id: &str, id: &str) -> UserLabelUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates the specified label.

fn threads_untrash(&self, user_id: &str, id: &str) -> UserThreadUntrashMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes the specified thread from the trash.

fn labels_patch(&self, request: &Label, user_id: &str, id: &str) -> UserLabelPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates the specified label. This method supports patch semantics.

fn drafts_delete(&self, user_id: &str, id: &str) -> UserDraftDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified draft. Does not simply trash it.

fn threads_list(&self, user_id: &str) -> UserThreadListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists the threads in the user's mailbox.

fn threads_modify(&self, request: &ModifyThreadRequest, user_id: &str, id: &str) -> UserThreadModifyMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Modifies the labels applied to the thread. This applies to all messages in the thread.

fn threads_delete(&self, user_id: &str, id: &str) -> UserThreadDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified thread. This operation cannot be undone. Prefer threads.trash instead.

fn messages_attachments_get(&self, user_id: &str, message_id: &str, id: &str) -> UserMessageAttachmentGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the specified message attachment.

fn messages_modify(&self, request: &ModifyMessageRequest, user_id: &str, id: &str) -> UserMessageModifyMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Modifies the labels on the specified message.

fn threads_trash(&self, user_id: &str, id: &str) -> UserThreadTrashMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Moves the specified thread to the trash.

fn drafts_list(&self, user_id: &str) -> UserDraftListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists the drafts in the user's mailbox.

fn messages_send(&self, request: &Message, user_id: &str) -> UserMessageSendMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sends the specified message to the recipients in the To, Cc, and Bcc headers.

fn messages_get(&self, user_id: &str, id: &str) -> UserMessageGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the specified message.

fn messages_list(&self, user_id: &str) -> UserMessageListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists the messages in the user's mailbox.

fn get_profile(&self, user_id: &str) -> UserGetProfileMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the current user's Gmail profile.

fn threads_get(&self, user_id: &str, id: &str) -> UserThreadGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the specified thread.

fn messages_insert(&self, request: &Message, user_id: &str) -> UserMessageInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Directly inserts a message into only this user's mailbox similar to IMAP APPEND, bypassing most scanning and classification. Does not send a message.